requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
pyTelegramBotAPI==4.2.0
python-dotenv==0.19.1
requests==2.26.0
urllib3==1.26.7
